+++ packages/react/src/index.ts
@@ -232,3 +232,22 @@ export {
   type UseNetworkOrdersParameters,
   type UseNetworkOrdersReturnType,
 } from './hooks/useNetworkOrders.js'
+
+
+export {
+  useBackOfQueueOrders,
+  type UseBackOfQueueOrdersParameters,
+  type UseBackOfQueueOrdersReturnType,
+} from './hooks/useBackOfQueueOrders.js'
+
+export {
+  useBackOfQueueBalances,
+  type UseBackOfQueueBalancesParameters,
+  type UseBackOfQueueBalancesReturnType,
+} from './hooks/useBackOfQueueBalances.js'
+
+export {
+  useBackOfQueueWallet,
+  type UseBackOfQueueWalletParameters,
+  type UseBackOfQueueWalletReturnType,
+} from './hooks/useBackOfQueueWallet.js'