+++ packages/react/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @renegade-fi/react
 
+## 0.0.34
+
+### Patch Changes
+
+- export hooks
+- Updated dependencies
+  - @renegade-fi/core@0.0.34
+
 ## 0.0.33
 
 ### Patch Changes