+++ packages/test/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @renegade-fi/test
 
+## 0.0.34
+
+### Patch Changes
+
+- Updated dependencies
+  - @renegade-fi/core@0.0.34
+
 ## 0.0.33
 
 ### Patch Changes